PXD-998 overrode styles for arranger table
