+++ src/css/base.less
@@ -17,6 +17,7 @@
 @white: #ffffff;
 @medium: 500;
 @semi-bold: 600;
+@body-font: 'Source Sans Pro', sans-serif;
 
 @-webkit-keyframes color-change {
   0% { color: @gray; }
@@ -25,7 +26,7 @@
 }
 
 html, body {
-  font-family: 'Source Sans Pro', sans-serif;
+  font-family: @body-font;
   margin: 0;
   display: flex;
   flex-direction: column;
