+++ src/stories/tables.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
 import 'regenerator-runtime/runtime';
-import DataExplorerTable from '../components/tables/DataExplorerTable';
+import DataExplorerTable from '../components/tables/DataExplorerTable/.';
 import { Table } from '@arranger/components/dist/DataTable';
 import '@arranger/components/public/themeStyles/beagle/beagle.css';
